Another try.
found the problem I've had with the spectra before: It's about the way
there are created. In order to increase the range of your spectrum, you
cannot simply increase the variables "ENumberChannel" and
"ENumberChannel".
I'll set this up as an issue and try to solve it, soon.

I guess it would be helpful to have a look on how the spectra are saved
in the examples.