Working - though not beautiful
It should now be working to create Energy histograms. They are produced
as root files in the data folder, named Edep.root. The file is
overwritten, so mv/cp it each time before you change the energy

- Again, think about the Scintillation part: (un)comment it here:
https://github.com/fzeiser/OCL/blob/master/src/SteppingAction.cc#L46-L50

- and(!), I already reduced the ScintYield to 0.008

There is a lot of things to be done more beatiful, but in principle it
works now.