Mix: Add GetPeaks for unfolding and efficiency+...
Add GetPeaks for an automatic analysis of the response function and for a table  of counts in the peaks. This can be used with Eff.py to plot the efficiency of OSCAR.

AnalysisSims has been moved to the data directory.

The macros runssims->genAndrunSims.sh and PrintSimMacs have been updated