CAD for LaBr3 housing
Implemented CAD model for the housing of the LaBr3 detectors. The files with the stainless steel part are commented out for the moment, as there had been some geometry overlap.
Needed also to change the ordering of the dauther/mother volumes; after implementing hte CAD drawings, initially it said that the housing fully encapsulates the detector on the same level. So I set the housing as the mother volume and implemented the detector inside.